workspace/slide_to_notion_converter/requirements.txt
notion-py
tkinter
python-pptx